+++ src/main/java/com/monew/monew_batch/job/api_article_collection/dto/NaverArticleResponse.java
@@ -1,0 +1,23 @@
+package com.monew.monew_batch.job.api_article_collection.dto;
+
+import java.util.List;
+
+import lombok.Data;
+
+@Data
+public class NaverArticleResponse {
+	String lastBuildDate;
+	int total;
+	int start;
+	int display;
+	List<ArticleItem> items;
+
+	@Data
+	public static class ArticleItem {
+		String title;
+		String originallink;
+		String link;
+		String description;
+		String pubDate;
+	}
+}
