--- src/main/java/com/monew/monew_batch/job/reader/YonhapArticleRssReader.java
@@ -1,51 +1,0 @@
-package com.monew.monew_batch.job.reader;
-
-import java.util.Iterator;
-import java.util.List;
-
-import org.springframework.batch.item.ExecutionContext;
-import org.springframework.batch.item.ItemStreamReader;
-import org.springframework.stereotype.Component;
-
-import com.monew.monew_batch.job.dto.ArticleSaveDto;
-import com.monew.monew_batch.job.reader.client.YonhapRssClient;
-import com.monew.monew_batch.properties.YonhapArticleProperties;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class YonhapArticleRssReader implements ItemStreamReader<ArticleSaveDto> {
-
-	private final YonhapRssClient yonhapRssClient;
-	private final YonhapArticleProperties yonhapArticleProperties;
-
-	private int feedIndex;
-	private List<String> feeds;
-	private Iterator<ArticleSaveDto> articleBufferIterator;
-
-	@Override
-	public void open(ExecutionContext executionContext) {
-		feedIndex = 0;
-		feeds = yonhapArticleProperties.getFeeds();
-	}
-
-	@Override
-	public ArticleSaveDto read() {
-		if (feedIndex >= feeds.size()) {
-			log.info("[연합뉴스] 모든 feed 대상 수집 완료");
-			feedIndex = 0;
-			return null;
-		}
-
-		if (articleBufferIterator == null || !articleBufferIterator.hasNext()) {
-			List<ArticleSaveDto> articleSaveDtos = yonhapRssClient.fetchArticles(feeds.get(feedIndex++));
-			log.info("[연합뉴스] 키워드 {}를 대상으로 수집 완료, 기사 개수 {}", feeds.get(feedIndex - 1), articleSaveDtos.size());
-			articleBufferIterator = articleSaveDtos.iterator();
-		}
-
-		return articleBufferIterator.next();
-	}
-}
